Add files via upload
The Clean Water is one of the rarest living fundamentals present on the earth. So, the measure should be taken to
conserve it for future purpose. Coming towards the farming which is the base of Indian providence the approaches used are still
the traditional. This leads to inappropriate use of water. So, to meet those conditions the end is to design the Smart Irrigation
System based on the detectors. This System would seek the purpose with the help of humidity, temperature detectors and
moisture detector that can be used for analyzing the environment. The humidity detectors and DHT11 detector will determine
the factual humidity status and temperature give the parameter to the microcontroller and hence the left work will be done on
those parameters by detectors. The total system will efficiently meet the water conditions by conserving water and will ease the
tradition irrigation system.